+++ src/main/java/net/sf/jabref/model/event/FieldChangedEvent.java
@@ -1,5 +1,6 @@
 package net.sf.jabref.model.event;
 
+import net.sf.jabref.event.source.EntryEventSource;
 import net.sf.jabref.model.FieldChange;
 import net.sf.jabref.model.entry.BibEntry;
 
@@ -15,6 +16,18 @@ public class FieldChangedEvent extends EntryChangedEvent {
      * @param bibEntry Affected BibEntry object
      * @param fieldName Name of field which has been changed
      * @param newValue new field value
+     * @param location location Location affected by this event
+     */
+    public FieldChangedEvent(BibEntry bibEntry, String fieldName, String newValue, EntryEventSource location) {
+        super(bibEntry, location);
+        this.fieldName = fieldName;
+        this.newValue = newValue;
+    }
+
+    /**
+     * @param bibEntry Affected BibEntry object
+     * @param fieldName Name of field which has been changed
+     * @param newValue new field value
      */
     public FieldChangedEvent(BibEntry bibEntry, String fieldName, String newValue) {
         super(bibEntry);
@@ -22,8 +35,20 @@ public class FieldChangedEvent extends EntryChangedEvent {
         this.newValue = newValue;
     }
 
+    /**
+     * @param bibEntry Affected BibEntry object
+     * @param fieldName Name of field which has been changed
+     * @param newValue new field value
+     * @param location location Location affected by this event
+     */
+    public FieldChangedEvent(FieldChange fieldChange, EntryEventSource location) {
+        super(fieldChange.getEntry(), location);
+        this.fieldName = fieldChange.getField();
+        this.newValue = fieldChange.getNewValue();
+    }
+
     public FieldChangedEvent(FieldChange fieldChange) {
-        this(fieldChange.getEntry(), fieldChange.getField(), fieldChange.getNewValue());
+        this(fieldChange, EntryEventSource.LOCAL);
     }
 
     public String getFieldName() {